fix: change ID data type from Integer to Long

+++ src/main/java/com/SmartBank/repository/AccountRepository.java
@@ -1,14 +1,13 @@
 package com.SmartBank.repository;
 
-import com.SmartBank.model.Account;
-import jakarta.validation.constraints.NotBlank;
+import com.SmartBank.entity.Account;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
 
 @Repository
-public interface AccountRepository extends JpaRepository<Account, Integer> {
+public interface AccountRepository extends JpaRepository<Account, Long> {
     boolean existsByAccountNumber(String accountNumber);
 
     List<Account> findAllByOrderByCustomerIdAsc();
