+++ src/main/java/com/SmartBank/repository/TransactionRepository.java
@@ -1,10 +1,10 @@
 package com.SmartBank.repository;
 
-import com.SmartBank.model.Transaction;
+import com.SmartBank.entity.Transaction;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TransactionRepository extends JpaRepository<Transaction, Integer> {
+public interface TransactionRepository extends JpaRepository<Transaction, Long> {
 
 }
